refactor: split workflow steps into separate modules
